+++ web/src/reducers/activities.js
@@ -8,13 +8,11 @@ export const setKeyGenerator = fn => {
 
 export const newGoal = () => ({
   key: generateKey(),
-  initialCollapsed: false,
   description: '',
   objective: ''
 });
 
 export const newMilestone = (milestone = '', endDate = '') => ({
-  initialCollapsed: false,
   key: generateKey(),
   milestone,
   endDate
@@ -23,7 +21,6 @@ export const newMilestone = (milestone = '', endDate = '') => ({
 export const statePersonDefaultYear = () => ({ amt: '', perc: '' });
 export const newStatePerson = years => ({
   key: generateKey(),
-  initialCollapsed: false,
   title: '',
   description: '',
   years: arrToObj(years, statePersonDefaultYear())
@@ -33,7 +30,6 @@ export const contractorDefaultYear = () => 0;
 export const contractorDefaultHourly = () => ({ hours: '', rate: '' });
 export const newContractor = years => ({
   key: generateKey(),
-  initialCollapsed: false,
   name: '',
   description: '',
   start: '',
@@ -51,7 +47,7 @@ export const expenseDefaultYear = () => 0;
 
 export const newExpense = years => ({
   key: generateKey(),
-  initialCollapsed: false,
+
   category: 'Hardware, software, and licensing',
   description: '',
   years: arrToObj(years, expenseDefaultYear())
@@ -92,7 +88,6 @@ export const newActivity = ({
   expenses: [newExpense(years)],
   fundingSource,
   goals: [newGoal()],
-  initialCollapsed: false,
   key: generateKey(),
   name,
   plannedEndDate: '',